Separate gguf reader test as standalone test matrix
